Remove test4.topo
The ibtopotool doesn't currently work with this topology file.
It's a bit difficult to reproduce this given the age, and IMO that
should be done first and foremost. The '#' on the last line breaks
formatting and theoretically that could even be just a copypaste
error from the document author or somesuch. Alternatively it could
be a very minimal PS1 prompt; hard to tell.